# openapi.yaml
openapi: 3.0.0
info:
  title: User Service
  description: Manages user accounts and profiles for the K11 platform
  version: 1.0.0
paths:
  /api/v2/users:
    get:
      summary: List all users
      operationId: list_users
      tags: [users]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user account
      operationId: create_user
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/v2/users/{id}:
    get:
      summary: Get a user by ID
      operationId: get_user
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
  /api/v2/users/{id}/contact:
    get:
      summary: Get a user's contact information
      operationId: get_user_contact
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Contact info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactInfo'
        '404':
          description: User not found
components:
  schemas:
    User:
      type: object
      required: [id, email, name]
      properties:
        id:
          type: string
          description: Unique user identifier
        email:
          type: string
          format: email
        name:
          type: string
        avatar_url:
          type: string
          nullable: true
    CreateUser:
      type: object
      required: [email, name]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
    ContactInfo:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true